+++ source/common/tcp_proxy/tcp_proxy.cc
@@ -111,10 +111,10 @@ Config::Config(const envoy::config::filter::network::tcp_proxy::v2::TcpProxy& co
 
 const std::string& Config::getRegularRouteFromEntries(Network::Connection& connection) {
   // First check if the per-connection state to see if we need to route to a pre-selected cluster
-  if (connection.streamInfo().perRequestState().hasData<PerConnectionCluster>(
+  if (connection.streamInfo().filterState().hasData<PerConnectionCluster>(
           PerConnectionCluster::Key)) {
     const PerConnectionCluster& per_connection_cluster =
-        connection.streamInfo().perRequestState().getData<PerConnectionCluster>(
+        connection.streamInfo().filterState().getData<PerConnectionCluster>(
             PerConnectionCluster::Key);
     return per_connection_cluster.value();
   }